requirements.txt
lxml==4.6.3
pyyaml==6.0.0
